+++ src/client/src/pages/dictionary/dictionary.ts
@@ -7,106 +7,22 @@ import {
 	dom,
 	hideElement,
 	notification,
-	partOfSpeechToClassName,
-	showElement,
 	validation,
 } from "~/shared/index.js";
 
-let audioSrc = "";
-
-const renderWordDefinition = (data: WordDto) => {
-	const partsOfSpeech = new Set(
-		data.meanings.map((meaning) => meaning.partOfSpeech),
-	);
-
-	showElement(".search-results__heading");
-	showElement(".word");
-	showElement(".transcription");
-	showElement(".definitions");
-
-	dom.setText({
-		selector: ".word",
-		text: data.word,
-	});
-
-	dom.setText({
-		selector: ".transcription-content",
-		text: data.phonetic?.text ?? "",
-	});
-
-	dom.clearContent(".badges");
-
-	partsOfSpeech.forEach((partOfSpeech) => {
-		dom.createElement({
-			className: ["badge", partOfSpeechToClassName[partOfSpeech]],
-			content: partOfSpeech,
-			parentElementSelector: ".badges",
-			tagName: "p",
-		});
-	});
-
-	if (data.phonetic?.audio) {
-		showElement(".play-transcription-btn");
-		audioSrc = data.phonetic.audio;
-	} else {
-		hideElement(".play-transcription-btn");
-	}
-
-	if (data.meanings) {
-		dom.clearContent(".definitions");
-		dom.createElement({
-			className: "definitions__header",
-			content: "Визначення",
-			parentElementSelector: ".definitions",
-			tagName: "h3",
-		});
-
-		dom.createElement({
-			className: "definitions-container",
-			parentElementSelector: ".definitions",
-			tagName: "div",
-		});
-
-		data.meanings.forEach((meaning, index) => {
-			dom.createElement({
-				className: ["definition", `definition--${index}`],
-				parentElementSelector: ".definitions-container",
-				tagName: "div",
-			});
-
-			dom.createElement({
-				className: "definition__word",
-				content: `${data.word} (${meaning.partOfSpeech})`,
-				parentElementSelector: `.definition--${index}`,
-				tagName: "p",
-			});
-
-			dom.createElement({
-				className: ["definition-list", `definition-list--${index}`],
-				parentElementSelector: `.definition--${index}`,
-				tagName: "ul",
-			});
-
-			meaning.definitions.forEach((definition) => {
-				const exampleParagraph = `<p class="example">Example: ${definition?.example}</p>`;
-
-				dom.createElement({
-					className: "definition-list-item",
-					content: `<p>${definition.definition}</p>${definition.example ? exampleParagraph : ""}`,
-					parentElementSelector: `.definition-list--${index}`,
-					tagName: "li",
-				});
-			});
-		});
-	}
-};
+import {
+	renderWordDefinition,
+	resetSearchInput,
+	showLoadingState,
+	showNotFound,
+	showSearchResults,
+} from "./libs/helpers/helpers.js";
+import { audio } from "./libs/modules/modules.js";
 
 const configureSearchForm = (): void => {
 	const handleFormSubmit = async () => {
 		const inputValue = dom.getElement<HTMLInputElement>("#search").value;
-		showElement(".loader-container");
-		hideElement(".search-results");
-		hideElement(".search-not-found");
+		showLoadingState();
 
 		try {
 			if (
@@ -124,27 +40,34 @@ const configureSearchForm = (): void => {
 			});
 
 			if (!("status" in data)) {
-				hideElement(".loader-container");
-				showElement(".search-results");
+				showSearchResults();
+				renderWordDefinition(data, (newValue: string) => {
+					audio.src = newValue;
+				});
 
-				renderWordDefinition(data);
-
-				dom.getElement<HTMLInputElement>("#search").value = "";
+				resetSearchInput();
 			} else if ("message" in data) {
-				hideElement(".loader-container");
-				showElement(".search-not-found");
+				showNotFound();
 				notification.error(data.message as string);
 
-				dom.getElement<HTMLInputElement>("#search").value = "";
+				resetSearchInput();
 			}
 		} catch (error) {
-			hideElement(".loader-container");
-			showElement(".search-not-found");
+			showNotFound();
+
 			if (error instanceof Error) {
 				notification.error(error.message);
 			}
 		}
 	};
+
+	const urlParams = new URLSearchParams(window.location.search);
+	const searchWord = urlParams.get("word");
+
+	if (searchWord) {
+		dom.getElement<HTMLInputElement>("#search").value = searchWord;
+		handleFormSubmit();
+	}
 
 	dom.setListener({
 		eventType: "submit",
@@ -177,8 +100,8 @@ const configureTranscriptionPlayer = (): void => {
 	dom.setListener({
 		eventType: "click",
 		listener: () => {
-			const audio = new Audio(audioSrc);
-			audio.play();
+			const audioObj = new Audio(audio.src);
+			audioObj.play();
 		},
 		selector: ".play-transcription-btn",
 	});