+++ src/client/src/pages/dictionary/libs/helpers/reset-search-input.helper.ts
@@ -1,0 +1,7 @@
+import { dom } from "~/shared/index.js";
+
+const resetSearchInput = () => {
+	dom.getElement<HTMLInputElement>("#search").value = "";
+};
+
+export { resetSearchInput };