+++ src/client/src/pages/dictionary/libs/helpers/show-not-found.helper.ts
@@ -1,0 +1,8 @@
+import { hideElement, showElement } from "~/shared/index.js";
+
+const showNotFound = () => {
+	hideElement(".loader-container");
+	showElement(".search-not-found");
+};
+
+export { showNotFound };