+++ src/client/src/pages/dictionary/libs/modules/modules.ts
@@ -1,0 +1,1 @@
+export { audio } from "./audio/audio.js";