+++ src/client/src/shared/libs/modules/dom/libs/types/types.ts
@@ -1,1 +1,2 @@
+export { type CreateElementOptions } from "./create-element-options.type.js";
 export { type DOM } from "./dom.type.js";